# Cache reflection calls when running utility lints and commands.