Reduces runtime of --check-yaml command to 70% of original. Will save time when running `make test` locally, and during CI runs.

| Timings                | Before | After | Relative |
| ------------------ | ------: | -----: | -------: |
| `ra --check-yaml` | 27.95s | 18.58s | 67% |
| `cnc --check-yaml` | 11.11s | 7.75s | 70% |
| `ts --check-yaml` | 9.28s | 6.26s | 68% |
| `d2k --check-yaml` | 6.63s | 4.54s | 69% |